+++ beamterm-data/src/atlas.rs
@@ -16,7 +16,7 @@ pub struct FontAtlasData {
     /// The font size in points
     pub font_size: f32,
     /// The number of single cell glyphs per layer (remaining glyphs are fullwidth)
-    pub halfwidth_glyphs_per_layer: u32,
+    pub halfwidth_glyphs_per_layer: u16,
     /// Width, height and depth of the texture in pixels
     pub texture_dimensions: (i32, i32, i32),
     /// Width and height of each character cell
